# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Healthtech Scheduling API",
        "description": "Agendamento de consultas: pacientes, profissionais e ciclo de vida da consulta.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Servidor atual"
        }
    ],
    "paths": {
...
        "/professionals": {
            "get": {
                "tags": [
                    "Professionals"
                ],
                "summary": "Lista profissionais",
                "operationId": "81bad49e55435925bc09a395da98311f",
                "responses": {
                    "200": {
                        "description": "Lista paginada de profissionais"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Professionals"
                ],
                "summary": "Cadastra um profissional com a janela de disponibilidade",
                "operationId": "d0a279eaafb72dedad02526abc07baa1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "specialty",
                                    "availabilities"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Dr. João"
                                    },
                                    "specialty": {
                                        "type": "string",
                                        "example": "Cardiologia"
                                    },
                                    "availabilities": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "weekday": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "start_time": {
                                                    "type": "string",
                                                    "example": "08:00"
                                                },
                                                "end_time": {
                                                    "type": "string",
                                                    "example": "18:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Profissional criado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/professionals/{professional}": {
            "get": {
                "tags": [
                    "Professionals"
                ],
                "summary": "Detalhe de um profissional",
                "operationId": "409ed0f209e762114805ce95a321d395",
                "parameters": [
                    {
                        "name": "professional",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profissional"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Professionals"
                ],
                "summary": "Atualiza um profissional (substitui a janela de disponibilidade)",
                "operationId": "4e9409c72f440e018a3fb69d6ded2232",
                "parameters": [
                    {
                        "name": "professional",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profissional atualizado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token obtido em POST /auth/login. Enviar como \"Authorization: Bearer {token}\".",
                "bearerFormat": "Sanctum token",
                "scheme": "bearer"
            }
        }
    },
...
}